# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "O-Ushers API",
    "description": "The **O-Ushers** platform connects event organizers with professional ushers (talents).\n\n## Roles\n| Role | Description |\n|------|-------------|\n| **Usher** | A talent who browses events, applies, and manages their profile |\n| **Organizer** | Posts events, manages applicants, marks attendance & reviews |\n| **Admin** | Platform-level management: users, events, verification |\n\n## Authentication\nAll protected endpoints require a **Bearer JWT** token obtained from `POST /auth/login`.\n\n## Pagination\nPaginated endpoints accept `page` and `size` query parameters and return:\n```json\n{ \"data\": [], \"currentPage\": 1, \"totalPages\": 5, \"totalCount\": 42 }\n```\n",
    "version": "1.0.0",
    "contact": {
      "email": "ooushers@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://o-ushers.vercel.app",
      "description": "Production server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
...
    "/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get own profile (any role)",
        "description": "Returns the authenticated user's full profile. Works for all roles.",
        "operationId": "getMyProfile",
        "responses": {
          "200": {
            "description": "Own profile",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/organizer/events": {
      "post": {
        "tags": [
          "Organizers"
        ],
        "summary": "Create event",
        "description": "Creates a new event. Status defaults to `open` and `hiredTalents` starts empty.",
        "operationId": "createEvent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventRequest"
              },
              "example": {
                "title": "Annual Corporate Gala",
                "category": "corporate",
                "eventDate": "2026-08-20T19:00:00Z",
                "applicationDeadline": "2026-08-10T23:59:59Z",
                "startTime": "19:00",
                "endTime": "23:00",
                "location": "Cairo Marriott Hotel",
                "requiredCount": 10,
                "genderPreference": "any",
                "budget": 5000,
                "dressCode": "Formal attire",
                "notes": "Arabic-speaking ushers preferred"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Event created",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Event"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "get": {
        "tags": [
          "Organizers"
        ],
        "summary": "Get own events",
        "description": "Returns a paginated list of the authenticated organizer's own events.",
        "operationId": "getOrganizerEvents",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "open",
                "confirmed",
                "completed",
                "cancelled"
              ]
            },
            "description": "Filter by event status."
          },
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/SizeParam"
          },
          {
            "$ref": "#/components/parameters/SortParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated events list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedEvents"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/organizer/events/{id}": {
      "get": {
        "tags": [
          "Organizers"
        ],
        "summary": "Get event detail",
        "description": "Returns full details of a single event owned by the organizer.",
        "operationId": "getOrganizerEventById",
        "parameters": [
          {
            "$ref": "#/components/parameters/EventIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Event detail",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Event"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "tags": [
          "Organizers"
        ],
        "summary": "Update event",
        "description": "Updates any of the allowed event fields. Only the organizer who owns the event\ncan update it.\n",
        "operationId": "updateEvent",
        "parameters": [
          {
            "$ref": "#/components/parameters/EventIdParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Event updated",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Event"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/organizer/events/{id}/close": {
      "patch": {
        "tags": [
          "Organizers"
        ],
        "summary": "Close event (confirm)",
        "description": "Sets the event status to `confirmed`, closing it for new applications.\n",
        "operationId": "closeEvent",
        "parameters": [
          {
            "$ref": "#/components/parameters/EventIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Event closed for new applications",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Event"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/organizer/events/{id}/applicants": {
      "get": {
        "tags": [
          "Organizers"
        ],
        "summary": "View event applicants",
        "description": "Returns all applications for a specific event with enriched talent info\nand referrer name.\n",
        "operationId": "getEventApplicants",
        "parameters": [
          {
            "$ref": "#/components/parameters/EventIdParam"
          }
        ],
        "responses": {
          "200": {
            "description": "List of applicants with talent details",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "allOf": [
                              {
                                "$ref": "#/components/schemas/Application"
                              },
                              {
                                "type": "object",
                                "properties": {
                                  "talent": {
                                    "$ref": "#/components/schemas/User"
                                  },
                                  "referredByName": {
                                    "type": "string",
                                    "nullable": true
                                  }
                                }
                              }
                            ]
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/organizer/events/{id}/attendance": {
      "post": {
        "tags": [
          "Organizers"
        ],
        "summary": "Mark attendance",
        "description": "Records or updates attendance for a hired talent. Business rules:\n- **FR-ATT-02**: Can only mark attendance after the application deadline or when event is confirmed.\n- Updates the talent's `consecutiveGoodEvents` counter (BR-07).\n- Resets the late-excuse counter after 5 consecutive good events (BR-06).\n- Triggers auto-verify check (FR-VER-01).\n",
        "operationId": "markAttendance",
        "parameters": [
          {
            "$ref": "#/components/parameters/EventIdParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "talentId",
                  "status"
                ],
                "properties": {
                  "talentId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "6ba45f12-1234-4562-c3ab-9f963f44bfa4"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "present",
                      "absent",
                      "late"
                    ],
                    "example": "present"
                  },
                  "checkInTime": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2026-08-20T19:15:00Z"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Attendance marked",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Attendance"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/organizer/events/{id}/reviews": {
      "post": {
        "tags": [
          "Organizers"
        ],
        "summary": "Review & rate talent",
        "description": "Submits a review and star rating (1–5) for a hired talent after the event.\nOne review per organizer per talent per event is allowed.\nAutomatically recalculates the talent's average `rate` and triggers\nauto-verify check (FR-VER-01).\n",
        "operationId": "reviewTalent",
        "parameters": [
          {
            "$ref": "#/components/parameters/EventIdParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "talentId",
                  "rating"
                ],
                "properties": {
                  "talentId": {
                    "type": "string",
                    "format": "uuid",
                    "example": "6ba45f12-1234-4562-c3ab-9f963f44bfa4"
                  },
                  "rating": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 5,
                    "example": 5
                  },
                  "comment": {
                    "type": "string",
                    "maxLength": 500,
                    "example": "Excellent performance, very professional!"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review submitted",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Review"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from `POST /auth/login`. Pass it in the\n`Authorization` header as `Bearer <token>`.\n"
      }
    },
    "parameters": {
      "PageParam": {
        "name": "page",
        "in": "query",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        },
        "description": "Page number (1-indexed)."
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Missing or invalid Bearer token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "message": "Authentication required"
            }
          }
        }
      },
...
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          }
        }
      },
...
    }
  }
}